requirements.txt
langchain-google-genai
langchain_community
tiktoken
google-generativeai
pypdf
python-dotenv
supabase
fastapi
uvicorn[standard]
pydantic
python-multipart
aiofiles
requests
replicate
pillow
pinecone
langchain-pinecone
langchain
boto3
tiktoken